backend/requirements.txt
fastapi
uvicorn
python-dotenv
SmartApi
pyotp
logzero
